requirements.txt
addict==2.4.0
einops==0.8.1
numpy==1.23.5
numpy==1.24.1
PyYAML==6.0.2
scikit_learn==1.2.2
scipy==1.15.3
timm==0.3.2
torch==1.13.1
